requirements.txt
tensorflow-cpu==2.16.1
flask
opencv-python
pillow
numpy==1.26.4
werkzeug
gunicorn
